% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gram_table.R
\name{create_gram_table}
\alias{create_gram_table}
\title{create_gram_table: Create n-gram tables with frequency - dataframe implementation}
\usage{
create_gram_table(x, n)
}
\description{
create_gram_table: Create n-gram tables with frequency - dataframe implementation
}
