% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gram_table.R
\name{create_gram_table2}
\alias{create_gram_table2}
\title{create_gram_table2: Create n-gram tables with frequency - data.table implementation}
\usage{
create_gram_table2(x, n)
}
\description{
create_gram_table2: Create n-gram tables with frequency - data.table implementation
}
