% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GT_discount.R
\name{GT_discount}
\alias{GT_discount}
\title{Get Good Turing discount from ngram frequency table}
\usage{
GT_discount(df, freq = r, mx = 6)
}
\arguments{
\item{df}{data.frame}

\item{freq}{name of freq column}

\item{mx}{max freq to consider for GT discount}
}
\description{
Get Good Turing discount from ngram frequency table
}
